+++ packages/core/src/index.ts
@@ -123,6 +123,7 @@ export { captureFeedback } from './feedback';
 export type { ReportDialogOptions } from './report-dialog';
 export { _INTERNAL_captureLog, _INTERNAL_flushLogsBuffer, _INTERNAL_captureSerializedLog } from './logs/exports';
 export { consoleLoggingIntegration } from './logs/console-integration';
+export { processVercelAiSpan } from './utils/vercel-ai';
 
 export type { FeatureFlag } from './utils/featureFlags';
 export {
