+++ packages/core/src/utils/vercel-ai-attributes.ts
